[expo-secure-store] Move iOS code to expo-secure-store
