+++ tools-public/package.json
@@ -34,7 +34,7 @@
     "sharp": "^0.21.0",
     "shelljs": "^0.7.0",
     "uuid": "^3.1.0",
-    "xdl": "52.0.14-beta.4"
+    "xdl": "52.1.1-beta.1"
   },
   "devDependencies": {
     "jest": "^23.5.0"
